Removes G10CurrencySelectionModel from Algorithm.Framework
Removes `G10CurrencySelectionModel` from `Algorithm.Framework` and implement them in the `G10CurrencySelectionModelFrameworkAlgorithm`
Minor change in the Python version of `ManualUniverseSelectionModel` to improve exception messages.
